install/djangoldap/requirements.txt
django
django-ldapdb
pytest
pytest-django
